+++ kwimage/structs/polygon.py
@@ -126,7 +126,7 @@ class _ShapelyMixin:
 
     def union(self, other):
         a, b = self.to_shapely(fix=1), other.to_shapely(fix=1)
-        c = a.intersection(b)
+        c = a.union(b)
         return _kwimage_from_shapely(c)
 
     def intersection(self, other):
@@ -1587,6 +1587,19 @@ class Polygon(_generic.Spatial, _PolyArrayBackend, _PolyWarpMixin, _ShapelyMixin
         xy = (shp_centroid.x, shp_centroid.y)
         return xy
 
+    def to_box(self):
+        """
+        Returns:
+            kwimage.Box
+        """
+        import kwimage
+        xys = self.data['exterior'].data
+        lt = xys.min(axis=0)
+        rb = xys.max(axis=0)
+        ltrb = np.hstack([lt, rb])
+        boxes = kwimage.Box.from_data(ltrb, 'ltrb')
+        return boxes
+
     def bounding_box(self):
         """
         Returns an axis-aligned bounding box for the segmentation
@@ -1883,6 +1896,10 @@ class Polygon(_generic.Spatial, _PolyArrayBackend, _PolyWarpMixin, _ShapelyMixin
             facecolor = color
         else:
             facecolor = kwimage.Color(facecolor, alpha=alpha)._forimage(image)
+
+        # TODO: consolidate logic
+        # _generic._handle_color_args_for(
+        #     color, alpha, border, fill, edgecolor, facecolor, image)
 
         if fill:
             if alpha is None or alpha == 1.0:
@@ -2362,6 +2379,22 @@ class MultiPolygon(_generic.ObjectList, _ShapelyMixin):
             kwimage.Boxes
         """
         return self.bounding_box()
+
+    def to_box(self):
+        """
+        Returns:
+            kwimage.Box
+        """
+        import kwimage
+        lt = np.array([np.inf, np.inf])
+        rb = np.array([-np.inf, -np.inf])
+        for data in self.data:
+            xys = data.data['exterior'].data
+            lt = np.minimum(lt, xys.min(axis=0))
+            rb = np.maximum(rb, xys.max(axis=0))
+        ltrb = np.hstack([lt, rb])
+        boxes = kwimage.Box.coerce(ltrb, format='ltrb')
+        return boxes
 
     def bounding_box(self):
         """